+++ lib/clixon/clixon_xml_io.h
@@ -45,7 +45,7 @@
  */
 int   clixon_xml2file1(FILE *f, cxobj *xn, int level, int pretty, char *prefix,
                        clicon_output_cb *fn, int skiptop, int autocliext, withdefaults_type wdef,
-                       int multi);
+                       int multi, int system_only);
 int   clixon_xml2file(FILE *f, cxobj *xn, int level, int pretty, char *prefix, clicon_output_cb *fn, int skiptop, int autocliext);
 int   clixon_xml2file_multi(clixon_handle h, const char *db, cxobj *xn, int level, int pretty,
                             char *prefix, clicon_output_cb *fn, int skiptop, int autocliext,